Add files via upload
Script can restart drivers that have names that contain variable $driverDescrip
Script uses the DeviceManagement module to restart drivers. Unlike the documentation which uses a pipeline to access the cmdlets Enable-Device and Disable-Device, this script passes a predefined variable to them which circumvents errors other programmers/scripters have had with this module and those cmdlets.